requirements.txt
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
ddgs>=1.15.0
exceptiongroup==1.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
Pillow>=10.0.0
primp==0.15.0
requests>=2.31.0
socksio==1.0.0
typing_extensions==4.15.0
urllib3==2.6.2

# ML / Training
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy>=1.27.0
tqdm>=4.66.1

# Backend / API (optional, for later)
fastapi>=0.111.0
uvicorn>=0.23.0
beautifulsoup4>=4.12.2
